Major refactor, bordering on rewrite
- Rename the plugin (md_yaml -> md_metayaml)
- Rework directory structure for easier integration into Pelican
- Change markdown extension libary to a submodule
- Remove ugly path hacks
- Make the reader do less work (leverage the MarkdownReader class)
