--- md_yaml/__init__.py
@@ -1,1 +1,0 @@
-from .md_yaml import *
